[FIX] Few MSVC warnings in the Matlab wrapper (comparison between signed and unsigned integers). [FIX] As the cameras' mask was removed in BTK 0.2, the functions SetPointData(Ex) display a warning. This warning was displayed in the Matlab tests.
git-svn-id: https://b-tk.googlecode.com/svn/BTK/trunk@1419 93be6e1e-6fcc-11de-baec-69240fdc299a

+++ Wrapping/Matlab/btkAppendMetaData.cpp
@@ -125,7 +125,7 @@ void mexFunction(int nlhs, mxArray *plhs[], int nrhs, const mxArray *prhs[])
     {
       if (!mxIsNumeric(temp) || mxIsEmpty(temp) || mxIsComplex(temp) || (mxGetNumberOfElements(temp) != 1))
         mexErrMsgTxt("The field 'numdims' must be set by a single integer value.");
-      int numDims = static_cast<int>(mxGetScalar(temp));
+      size_t numDims = static_cast<size_t>(mxGetScalar(temp));
       if (numDims < dims.size())
         mexErrMsgTxt("The given number of dimensions is lower than the number of dimensions for the given values.");
       else if (numDims > maxDimSize)
